fix(category-sphere): change colors for the category icons and update imports
