+++ src/components/category-sphere/index.ts
@@ -1,0 +1,1 @@
+export * from './category-icons'
